+++ plugins/postcss-global-data/.tape.mjs
@@ -1,6 +1,7 @@
 import { postcssTape } from '@csstools/postcss-tape';
 import plugin from '@csstools/postcss-global-data';
 import postcssCustomMedia from 'postcss-custom-media';
+import postcssCustomProperties from 'postcss-custom-properties';
 
 postcssTape(plugin)({
 	basic: {
@@ -12,6 +13,19 @@ postcssTape(plugin)({
 				]
 			}),
 			postcssCustomMedia(),
+		],
+	},
+	'open-props': {
+		message: "supports open-props usage",
+		plugins: [
+			plugin({
+				files: [
+					'node_modules://open-props/media.min.css',
+					'node_modules://open-props/open-props.min.css',
+				]
+			}),
+			postcssCustomProperties({ preserve: false }),
+			postcssCustomMedia({ preserve: false }),
 		],
 	},
 });